+++ config/initializers/scheduler.rb
@@ -1,4 +1,0 @@
-# if Rails.env.production?
-#   require './lib/scheduler.rb'
-#   Publication::FeedScheduler.new.run!
-# end