+++ db/migrate/20121019135911_remove_expires_on_from_items.rb
@@ -1,0 +1,9 @@
+class RemoveExpiresOnFromItems < ActiveRecord::Migration
+  def up
+    remove_column :items, :expires_on
+  end
+
+  def down
+    add_column :items, :expires_on, :datetime
+  end
+end