+++ db/migrate/20121115041908_add_twitter_hashtags_to_items.rb
@@ -1,0 +1,5 @@
+class AddTwitterHashtagsToItems < ActiveRecord::Migration
+  def change
+    add_column :items, :hashtags, :string
+  end
+end